+++ src/types/restaurantType.ts
@@ -1,0 +1,28 @@
+export interface RestaurantInfoType {
+  ADDR1 : string;
+  ADDR2 : string;
+  CNTCT_TEL : string;
+  GUGUN_NM : string;
+  HOMEPAGE_URL : string;
+  ITEMCNTNTS : string;
+  LAT : number;
+  LNG : number;
+  MAIN_IMG_NORMAL : string;
+  MAIN_IMG_THUMB : string;
+  MAIN_TITLE : string;
+  PLACE : string;
+  RPRSNTV_MENU : string;
+  SUBTITLE : string;
+  TITLE : string;
+  UC_SEQ : number;
+  USAGE_DAY_WEEK_AND_TIME : string;
+}
+
+
+export interface RestaurantInitialStateType {
+  status : null | string;
+  restaurantList : RestaurantInfoType[];
+  filteredRestaurant : RestaurantInfoType[];
+  optionRestaurant : string[];
+  currentFilter : string;
+}
