+++ src/types/wishToGoListType.ts
@@ -1,0 +1,14 @@
+export interface WishToGoListType{
+  ADDR1 : string;
+  GUGUN_NM : string;
+  ITEMCNTNTS : string;
+  LAT : number;
+  LNG : number;
+  MAIN_IMG_THUMB : string;
+  MAIN_TITLE : string;
+  TITLE : string;
+  UC_SEQ : number;
+  docId : string
+}
+
+
